Remove unneeded configuration dsl & methods
javascript_file & stylesheet_file are obsolete because lotus templates
methods can be called with params.
Also the javascript_engine & stylesheet_engine are obsolete because we
are no longer setting a fixed javascript_engine or stylesheet_engine.

Also renamed javascript_include_tag and stylesheet_include_tag to
javascript and stylesheet.

This means the user can say following:

    <%= stylesheet 'application' %>

And in the current assets path a file with the name application.scss is
picked up automatically.