Configuration class should also set configs
Currently the Dsl class was responsible to set the Configuration
attributes via proper attr_writer. The Configuration class should use
the same method for setting and reading the config (if nil -> read it).

This makes the logic in the Dsl class obsolete and Dsl is only a
delegator to the current Configuration class attribute.