Unified asset sources.
On before each asset type (javascript, stylesheet) used to have its own
definition of sources. That was repetitive to configure and didn't
matched real use cases.

We went from:

  Lotus::Assets.configure do
    # ...
    define(:javascript) do
      sources << [
        'assets',
        'vendor/assets/bootstrap/javascripts'
      ]
    end

    define(:stylesheet) do
      sources << [
        'assets',
        'vendor/assets/bootstrap/stylesheets'
      ]
    end
  end

To this:

  Lotus::Assets.configure do
    sources << [
      'assets',
      'vendor/assets/bootstrap'
    ]
  end

This DRY configuration has another win: it lets devs to load all the
files defined under a source, without worrying about the asset type.
For instance Bootstrap ships with fonts and images that we want to
serve without the need of manually configure this kind of assets.