Compare timestamps instead of dates
When comparing dates (on open file, I presume) the comparison with <=
shows that dates are different, even though difference is on sub-second
level. As a result, when comparing later, mtimes of both times are the
same, leading at least tests to fail.
The workaround is to cast both dates to timestamp using #to_i before comparison.