Compare timestamps instead of dates
When comparing dates (on open file, I presume) the comparison with <=
shows that dates are different, even though difference is on sub-second
level. As a result, when comparing later, mtimes of both times are the
same, leading at least tests to fail.
The workaround is to cast both dates to timestamp using #to_i before comparison.

This change was already applied in commit 230f3265473096da5e4c6bf6ea5836106e512883
but was lost while tests were rewritten into RSpec.