Update nested asset spec to test correct behavior
This spec wasn't being run by CI, so it wasn't noticed that the spec
itself was testing the incorrect behavior.

The behavior of the production code is correct, which was demonstrated
and verified in https://github.com/hanami/asset/issues/95. This commit
alters the test to reflect the behavior of the production code.