Compile assets only once; Explicitly target syntax
In the path to replacing Sass with SassC (libsass), we've been working
to ensure the functionality is equivalent. SassC doesn't rely on
caching, but we cache our dependencies. In order for us to cache our
dependencies, we need to get them from the renderer.

By memoizing the renderer, we'll either render the asset _then_ allow
access to its dependencies (for caching) OR we'll bubble up the
underlying exception that SaccC throws when dependencies have yet to be
rendered.