Wrap SassC/Sass gems
This is in response to a bug which was pointed out in
https://github.com/hanami/assets/pull/106. We added `require 'sassc'`,
which turned out to be a breaking change. If a user doesn't have `sassc`
listed in their Gemfile, they'll get load errors.

In order to resolve this, I'm opting to wrap both `SassC` and `Sass`,
with the idea that we can remove the reliance on `Sass` soon (the
sass-ruby gem is past end-of-life).