Introduce Hanami::Assets::Source
This allows us to extract the manifest from Configuration, and let Configuration remain a much simpler object responsible for holding the settings only, and not returning any “active” components of Hanami::Assets.

I’m not 100% happy with the name “Source”, but I consider this a reasonable next step as we figure out the best structure for this gem.