Introduce Hanami::Assets::Asset
This can represent a single asset and provide a rich set of methods for it. Start with methods for the path, url, and subresource_integrity_value.

Return asset instances from Source#[], and correspondingly remove the separate Source#asset_path and #subresource_integrity_value methods.

There’s still a naming tension yet to be resolved here, in that Helpers#path uses what is actually the Asset#url. Ideally Helpers#path would use Asset#path underneath (or maybe we need a set of methods on Helpers, one for `_url`, one for `_path`).