Rewrite for 2.1 and esbuild-based hanami-assets-js package (#120)
* Reset the specs

* Enable RuboCop

* Rewrite `Hanami::Assets::Precompiler`

* Introduced `Hanami::Assets::Watcher`

* Refactoring: use `esbuild` via build script.

* Generate `manifest.json` when precompiling

* Ensure to use `esbuild-plugin-manifest` only when precompiling.

Ensure unsucessful process exit when precompile fails.

* Setup CI with Node/Yarn

* Revert "Setup CI with Node/Yarn"

This reverts commit dc4bc79c19eb2ad48562864c7b12eb9e9a9027fa.

Move `ravsamhq/notify-slack-action@v1` at the bottom

* CI: setup Node

* CI: Yarn install

* Remove `yarn.lock` from git

* Integration test to use `Rack::Files` to serve assets statically

* Point to `hanami-devtools` `main` Git branch

* More permissive assertion

* Add .repobot.yml config

* Enable integration specs

* Upgrade from `esbuild` `"^0.14.23"` to `"^0.17.06"`

* Test against React 18.2

* Assets minification during precompilation

* Ensure serve_spec.rb to simulate real world development experience

* Introducing `Hanami::Assets::Helpers#[]` to return relative/absolute path to asset

* Use `hanami-esbuild` plugin for `esbuild`

* Add extra test for `Helpers#[]`

* Re-add tests for `Hanami::Assets::Helpers#javascript`

* Adapt to the latest changes of `hanami-esbuild`

* Support Subresource Integrity

* Adapt to latest `hanami-esbuild` changes. Support for `#stylesheet` helper.

* Support for `#image` helper

* Support for `#favicon` helper

* Support for `#video` helper

* Support for `#audio` helper

* Support for `#asset_path` helper

* Rationalize `#asset_path` behavior. Removed `#asset_url`.

* Update GitHub Actions CI workflow

* Depend on `main` `hanami-view` branch

* Revert "Depend on `main` `hanami-view` branch"

This reverts commit c3a543c8ed39746bbacf29a5c298f299826f4ccf.

* Remove no longer needed code

* Install `hanami-esbuild` NPM package via GitHub

* Use `npm` instead of `yarn` in CI

* Depend on `hanami-view` at `main` branch

* Remove unnecessary `require` statements

* Use `dry-inflector` instead of `Hanami::Utils::String`

* Cleanup

* README

* Enable Rubocop

* Remove unused fixtures

* Simplify RuboCop configuration

* Remove `script` directory

* CHANGELOG

* Use `hanami-assets` NPM package from `hanami/assets-js`

* Fix README

* Detect JS executable during runtime, not at boot time

* Add `sources` and `destination` kwargs to `Hanami::Assets::Configuration`

* Update deps (2023-07-26)
Ref https://github.com/hanami/assets-js/commit/de434397ae5dc1e297afe40bcba252c1cd5b3cc9

* Updated Ruby deps

* Update Configuration to use dry-configurable

* Introduce Hanami::Assets::Source

This allows us to extract the manifest from Configuration, and let Configuration remain a much simpler object responsible for holding the settings only, and not returning any “active” components of Hanami::Assets.

I’m not 100% happy with the name “Source”, but I consider this a reasonable next step as we figure out the best structure for this gem.

* Introduce Hanami::Assets::Asset

This can represent a single asset and provide a rich set of methods for it. Start with methods for the path, url, and subresource_integrity_value.

Return asset instances from Source#[], and correspondingly remove the separate Source#asset_path and #subresource_integrity_value methods.

There’s still a naming tension yet to be resolved here, in that Helpers#path uses what is actually the Asset#url. Ideally Helpers#path would use Asset#path underneath (or maybe we need a set of methods on Helpers, one for `_url`, one for `_path`).

* Remove stale manifest references in Configuration

* Rename Assets::Source to just Assets

Assets::Source always felt awkward, especially given we had to pass `source_path` arguments to its methods. In the end, I couldn’t think of any better name for this class (whose role is to provide access to the assets themselves) than just straight up `Assets`. Usage now feels much more natural.

* Simplify helpers by removing push promise support

AFAICT, we don’t expose push/promise support in Hanami 1, nor do we have plans to do so in Hanami 2, nor does any other major Ruby web framework, so we do ourselves a favour for now by removing this and greatly simplifying our helpers.

* Remove need for configuration to be given to helpers

* Add _tag suffixes to helper names

This makes it clearer what they’re returning.

* Remove #[] alias from helpers

If we make this a module, this is too ambiguous to be helpful

* Remove precompiler from specs that don’t need it

* Add missing require

* Turn helpers into a module expecting a hanami-view _context

* Rename #path helper to #asset_url

* Shift reader up

* Rename Configuration to Config

* Remove unneeded super

* Tidy spacing

* Gracefully handle a missing manifest, even if manifest_path is configured

* Remove helpers

I’ve moved these to the hanami gem

* Avoid passing whole config to Asset objects

---------

Co-authored-by: Peter Solnica <peter@solnica.online>
Co-authored-by: Tim Riley <tim@riley.id.au>